+++ CHANGELOG.md
@@ -4,6 +4,7 @@ PowerModels.jl Change Log
 ### Staged
 - Add switch component (#339)
 - Added component status parameter lookup (#565)
+- Relax JuMP Model type (#582)
 - Fixed status-based filtering to support fractional values (#564)
 - Fixed `va` default start point